Add a script to automatically install git hooks
Since meson depends on python 3.6+ anyway, we can use python to make a
cross-platform git-hook installer.

The script is run by meson automatically, but also works when run
manually.

It will scan the directory it is located in for files ending with
".hook" - any files found will be installed into ".git/hooks" with the
".hook" suffix removed and permissions set to 755.